feat: InsightFace识别、模型管理、首次配置向导、GUI优化
- 新增 InsightFace ArcFace-R100 识别器 (512维, 预训练高精度)
- 新增模型管理对话框: 一键下载buffalo_l, 暂停/恢复/取消, 断点续传
- 新增首次运行配置向导: 自动检测缺失组件并引导安装
- 新增编码按模型隔离存储 (embeddings_mbn/if/fr.bin)
- 新增工作照上传裁剪功能
- GUI自适应缩放 (resizeEvent动态调整字体/间距/行高)
- 表格整行hover/选中高亮 (RowHighlightDelegate)
- C++部署: config.json配置 + 模型文件自动搜索
- 清理冗余文件: train.py, train_simple.py, test_model.py等
- 修复sqlite_db.py float64/float32不一致bug
- 更新README和子模块文档

+++ common/databases/sqlite_db.py
@@ -83,7 +83,7 @@ class SQLiteDatabase(FaceDatabase):
 
             if rows:
                 # 计算平均 embedding
-                encodings = [np.frombuffer(row[0], dtype=np.float64) for row in rows]
+                encodings = [np.frombuffer(row[0], dtype=np.float32) for row in rows]
                 avg_encoding = np.mean(encodings, axis=0)
                 result[name] = avg_encoding
 
@@ -114,7 +114,7 @@ class SQLiteDatabase(FaceDatabase):
                 user_id = cursor.lastrowid
 
             # 保存编码
-            encoding_bytes = embedding.astype(np.float64).tobytes()
+            encoding_bytes = embedding.astype(np.float32).tobytes()
             cursor.execute(
                 "INSERT INTO face_encodings (user_id, encoding) VALUES (?, ?)",
                 (user_id, encoding_bytes)
